+++ internal/pkg/constants/dag_actions.go
@@ -2,6 +2,7 @@ package constants
 
 const DagActionInstall = "install"
 const DagActionDelete = "delete"
+const DagActionClean = "clean"
 const DagActionTemplate = "template"
 const PostActionClusterUpdate = "cluster_update"
 const PostActionAddProviderVarsFiles = "add_provider_vars_files"