+++ internal/pkg/interfaces/installer.go
@@ -19,6 +19,7 @@ package interfaces
 type IInstaller interface {
 	Install(installableObj IInstallable, stack IStack, approved bool, dryRun bool) error
 	Delete(installableObj IInstallable, stack IStack, approved bool, dryRun bool) error
+	Clean(installableObj IInstallable, stack IStack, dryRun bool) error
 	Output(installableObj IInstallable, stack IStack, dryRun bool) error
 	Name() string
 	GetVars(action string, approved bool) map[string]interface{}